put team3's experiment in a separate folder
