+++ textdb/textdb-sandbox/src/test/java/edu/uci/ics/textdb/sandbox/team3/team3lucenecityexample/LuceneDemoTest.java
@@ -1,9 +1,9 @@
-package edu.uci.ics.textdb.sandbox.team3lucenecityexample;
+package edu.uci.ics.textdb.sandbox.team3.team3lucenecityexample;
 
-import static edu.uci.ics.textdb.sandbox.team3lucenecityexample.LuceneIndexConstants.CONTENT_FIELD;
-import static edu.uci.ics.textdb.sandbox.team3lucenecityexample.LuceneIndexConstants.ID_FIELD;
-import static edu.uci.ics.textdb.sandbox.team3lucenecityexample.LuceneIndexConstants.NAME_FIELD;
-import static edu.uci.ics.textdb.sandbox.team3lucenecityexample.LuceneIndexConstants.COUNTRY_FIELD;
+import static edu.uci.ics.textdb.sandbox.team3.team3lucenecityexample.LuceneIndexConstants.CONTENT_FIELD;
+import static edu.uci.ics.textdb.sandbox.team3.team3lucenecityexample.LuceneIndexConstants.ID_FIELD;
+import static edu.uci.ics.textdb.sandbox.team3.team3lucenecityexample.LuceneIndexConstants.NAME_FIELD;
+import static edu.uci.ics.textdb.sandbox.team3.team3lucenecityexample.LuceneIndexConstants.COUNTRY_FIELD;
 
 import java.io.IOException;
